trying to move all the datetime utils and logic to a ginslge file
