+++ lmfdb/lmfdb_database.py
@@ -90,7 +90,8 @@ class LMFDBSearchTable(PostgresSearchTable):
         - ``other_table`` -- a string, the name of the other table.
         - ``keep_old`` -- if true, new knowls for this table will be created from the column knowls for the old table.  Otherwise, the old knowls will be renamed, or deleted if they are not columns of this table.
         """
-        from lmfdb.knowledge.knowl import knowldb, utc_now_naive
+        from lmfdb.knowledge.knowl import knowldb
+        from lmfdb.utils.datetime import utc_now_naive
         knowls = knowldb.get_column_description(other_table)
         with DelayCommit(self):
             for col, knowl in knowls.items():
